Move DepthUnprojection to gfx_batch.
